Removed all references to the serverside analysis pipeline
