--- src/analysis/offline-transformers/map.js
@@ -1,14 +1,0 @@
-/**
- * map a key at path (eg. "climb1") to a value (eg. 4)
- * @type {DataTransformer}
- * @param options.path {String} the path to which the key is found
- * @param options.map {Object} an Object that contains the keys and values. If the path to the 
- */
-/* <TMP> */
-new DataTransformer("map", (dataset, outputPath, options) => {
-    for (let tmp of dataset.tmps) {
-        setPath(tmp, outputPath, options.map[getPath(tmp, options.path,"")]) //default to an empty string if there is no value
-    }
-    return dataset;
-})
-/* </TMP> */
