--- src/analysis/offline-transformers/subtract.js
@@ -1,32 +1,0 @@
-/**
- * @type {DataTransformer}
- * @param options.minuend {String} a MatchTeamPerformance outputPath with the minuend
- * @param options.subtrahend {String} a MatchTeamPerformance outputPath with the subtrahend
- */
-/* <TMP> */
-new DataTransformer("subtract", (dataset, outputPath, options) => {
-    for (const tmp of dataset.tmps) {
-        const difference = getPath(tmp, options.minuend, 0) - getPath(tmp, options.subtrahend, 0)
-
-        setPath(tmp, outputPath, difference)
-    }
-
-    return dataset;
-})
-/* </TMP> */
-
-/**
- * @param options.minuend {String} a MatchTeamPerformance outputPath with the minuend
- * @param options.subtrahend {String} a MatchTeamPerformance outputPath with the subtrahend
- */
-/* <TEAM> */
-new DataTransformer("subtract", (dataset, outputPath, options) => {
-    for (const [teamNumber, team] of Object.entries(dataset.teams)) {
-        const difference = getPath(team, options.minuend, 0) - getPath(team, options.subtrahend, 0)
-
-        setPath(team, outputPath, difference)
-    }
-
-    return dataset;
-})
-/* </TEAM> */
