--- src/analysis/offline-transformers/template.js
@@ -1,19 +1,0 @@
-/**
- * @type {DataTransformer}
- * @param options.example {String} example parameter description
- */
-/* <TMP> */
-new DataTransformer("name", (dataset, outputPath, options) => {
-    return dataset;
-})
-/* </TMP> */
-
-/**
- * @type {DataTransformer}
- * @param options.example {String} example parameter description
- */
-/* <TEAM> */
-new DataTransformer("name", (dataset, outputPath, options) => {
-    return dataset;
-})
-/* </TEAM> */
