+++ src/analysis/routes/api.js
@@ -1,19 +1,13 @@
 const { Router } = require("express");
-const { execute, transformers, manual } = require("../analysisPipeline.js")
 const { TeamMatchPerformance } = require('../../lib/db.js');
  const axios = require("axios")
 const config = require("../../../config/config.json");
-const { DataTransformer } = require("../DataTransformer.js");
 const {getTempTeams} = require("../../schedule/schedule");
 
 let router = Router();
 
 router.get("/dataset", async (req, res) => {
-    res.json(await execute())
-});
-
-router.get("/rawDataset", async (req, res) => {
-  res.json(await TeamMatchPerformance.find({eventNumber: config.EVENT_NUMBER}));
+    res.json(await TeamMatchPerformance.find({eventNumber: config.EVENT_NUMBER}));
 });
 
 if (!config.secrets.TBA_API_KEY) {
@@ -30,35 +24,15 @@ router.get("/teams", async (req, res) => {
         }
     }).catch(e => console.log(e,chalk.bold.red("\nError fetching teams from Blue Alliance API!")))).data;
     res.json(tbaTeams)
-})
-
-router.get("/transformers", async (req, res) => {
-  const cleanedTransformers = {};
-  const tmp = {};
-  const team = {};
-
-  const tmpTransformers = Object.keys(transformers.tmp);
-
-  for (const transformer of tmpTransformers) {
-    tmp[transformer] = {};
-    tmp[transformer].execute = transformers.tmp[transformer].execFunc.toString();
-  }
-
-  const teamTransformers = Object.keys(transformers.team);
-
-  for (const transformer of teamTransformers) {
-    team[transformer] = {};
-    team[transformer].execute = transformers.team[transformer].execFunc.toString();
-  }
-
-  cleanedTransformers.tmp = tmp;
-  cleanedTransformers.team = team;
-
-  res.json(cleanedTransformers);
 });
 
 router.get("/manual", async (req, res) => {
-  res.json(manual);
+    const manual = {
+        teams: require("../manual/teams.json"),
+        tmps: require("../manual/tmps.json")
+    }
+
+    res.json(manual);
 });
 
 router.get("/csv", async (req,res) => {
