+++ src/analysis/transformers/template.js
@@ -1,19 +1,19 @@
-const {DataTransformer} = require("../DataTransformer");
+/**
+ * @type {DataTransformer}
+ * @param options.example {String} example parameter description
+ */
+/* <TMP> */
+new DataTransformer("name", (dataset, outputPath, options) => {
+    return dataset;
+})
+/* </TMP> */
 
-module.exports = {
-    /**
-     * @type {DataTransformer}
-     * @param options.example {String} example parameter description
-     */
-    tmp: new DataTransformer("name", (dataset, outputPath, options) => {
-        return dataset;
-    }),
-
-    /**
-     * @type {DataTransformer}
-     * @param options.example {String} example parameter description
-     */
-    team: new DataTransformer("name", (dataset, outputPath, options) => {
-        return dataset;
-    })
-}
+/**
+ * @type {DataTransformer}
+ * @param options.example {String} example parameter description
+ */
+/* <TEAM> */
+new DataTransformer("name", (dataset, outputPath, options) => {
+    return dataset;
+})
+/* </TEAM> */
