+++ src/analysis/views/index.ejs
@@ -151,7 +151,7 @@
 	<script src="https://cdn.plot.ly/plotly-2.8.3.min.js"></script>
 	<script src="./js/DataTransformer.js"></script>
 	<script src="./modules.js"></script>
-	<script src="./js/offlinePipeline.js"></script>
+	<script src="./js/analysisPipeline.js"></script>
 	<script src="./transformers.js"></script>
 	<script src="./js/script.js"></script>
